avatar: Do not load if size is 0
Fixes a crash in big rooms because all avatars are loaded at once

+++ src/components/avatar.rs
@@ -138,7 +138,7 @@ impl Avatar {
     fn request_custom_avatar(&self) {
         if let Some(item) = &*self.imp().item.borrow() {
             // FIXME: update on size changes
-            item.set_needed_size(self.size());
+            item.set_needed_size(self.size() as u32);
         }
     }
 }
